src/main/java/ru/job4j/collection/User.java
package ru.job4j.collection;

import java.util.Objects;

/**
 * 1. Организовать сортировку User [#285695]
 */
public class User implements Comparable<User> {
    private String name;
    private int age;

    public User(String name, int age) {
        this.name = name;
        this.age = age;
    }

    @Override
    public int compareTo(User other) {
        int result = name.compareTo(other.name);
        if (result != 0) {
            return result;
        }
        result = age - other.age;
        return result;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        User user = (User) o;
        return age == user.age &&
                Objects.equals(name, user.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name, age);
    }
}
